Moved new component to new_designs folder
